# robin/openapi.yaml
openapi: "3.1.0"
info:
  title: Robin API
  version: "0.3.0"
  description: REST API for Robin.OS — a personal knowledge management system.
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - cookieAuth: []
...
paths:
...
  /vaults:
    get:
      operationId: listVaults
      summary: List all vaults
      tags: [Vaults]
      responses:
        "200":
          description: List of vaults
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/vaultListResponseSchema"
    post:
      operationId: createVault
      summary: Create a new vault
      tags: [Vaults]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createVaultBodySchema"
      responses:
        "201":
          description: Created vault
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/vaultResponseSchema"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
  /vaults/{id}:
    get:
      operationId: getVault
      summary: Get a vault by ID
      tags: [Vaults]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: The vault
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/vaultResponseSchema"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
    put:
      operationId: updateVault
      summary: Update a vault
      tags: [Vaults]
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/updateVaultBodySchema"
      responses:
        "200":
          description: Updated vault
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/vaultResponseSchema"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
  /vaults/{id}/profile:
    put:
      operationId: updateVaultProfile
      summary: Update vault profile text
      tags: [Vaults]
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Updated vault
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/vaultResponseSchema"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
  /vaults/{vaultId}/threads:
    get:
      operationId: listVaultThreads
      summary: List threads in a vault
      tags: [Vaults]
      parameters:
        - $ref: "#/components/parameters/VaultId"
      responses:
        "200":
          description: List of threads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/threadListResponseSchema"
    post:
      operationId: createVaultThread
      summary: Create a thread in a vault
      tags: [Vaults]
      parameters:
        - $ref: "#/components/parameters/VaultId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createThreadBodySchema"
      responses:
        "201":
          description: Created thread
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/threadResponseSchema"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"
        "404":
          description: Vault not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errorResponseSchema"

  # ── People ─────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie set by better-auth after sign-in
    hmacAuth:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC-SHA256 signature of request body
    mcpToken:
      type: apiKey
      in: query
      name: token
      description: JWT token for MCP access

  # ── Shared Parameters ──────────────────────────────────────────────
  parameters:
    Id:
      name: id
      in: path
      required: true
      description: Resource ID or lookup key
      schema:
        type: string
    VaultId:
      name: vaultId
      in: path
      required: true
      description: Vault ID
      schema:
        type: string

  # ── Standard Responses ─────────────────────────────────────────────
...
  schemas:
...
    errorResponseSchema:
      type: object
      properties:
        error:
          type: string
        fields: {}
      required: [error]
      additionalProperties: false
...
    createThreadBodySchema:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        type:
          type: string
          default: log
        prompt:
          type: string
      required: [name]
      additionalProperties: false
...
    threadResponseSchema:
      type: object
      properties:
        id:
          type: string
          minLength: 1
        lookupKey:
          type: string
          minLength: 1
        userId:
          type: string
        slug:
          type: string
        name:
          type: string
        type:
          type: string
        prompt:
          type: string
        state:
          type: string
          enum: [PENDING, RESOLVED, LINKING, DIRTY]
        repoPath:
          type: string
        vaultId:
          type: string
          nullable: true
        lastRebuiltAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        noteCount:
          type: number
          default: 0
        lastUpdated:
          type: string
      required: [id, lookupKey, userId, slug, name, type, prompt, state, repoPath, vaultId, lastRebuiltAt, createdAt, updatedAt, lastUpdated]
      additionalProperties: false
...
    threadListResponseSchema:
      type: object
      properties:
        threads:
          type: array
          items:
            $ref: "#/components/schemas/threadResponseSchema"
      required: [threads]
      additionalProperties: false

    # ── People ─────────────────────────────────────────────────────
...
    createVaultBodySchema:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        icon:
          type: string
        description:
          type: string
        color:
          type: string
      required: [name]
      additionalProperties: false
    updateVaultBodySchema:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        description:
          type: string
        profile:
          type: string
        color:
          type: string
      additionalProperties: false
    vaultResponseSchema:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        slug:
          type: string
        icon:
          type: string
        description:
          type: string
        profile:
          type: string
          nullable: true
        color:
          type: string
        type:
          type: string
          enum: [user, system, inbox]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        threadCount:
          type: number
          default: 0
        noteCount:
          type: number
          default: 0
      required: [id, userId, name, slug, icon, description, profile, color, type, createdAt, updatedAt]
      additionalProperties: false
    vaultListResponseSchema:
      type: object
      properties:
        vaults:
          type: array
          items:
            $ref: "#/components/schemas/vaultResponseSchema"
      required: [vaults]
      additionalProperties: false

    # ── Search ─────────────────────────────────────────────────────
...